Backend

+++ backend/event/views.py
@@ -1,25 +1,16 @@
 from django.shortcuts import render
 from django.http import HttpResponse
 from django.views.decorators.csrf import get_token
+from django.contrib.auth.decorators import login_required
 import json
 
 from .models import Event,EventForm
+from organization.models import Organization
 
-'''
-# API Syntax 
-## Events
-* /events/
-	* sortBy
-	* count
-* /events/delete
-	* id
-
-'''
-
-def events(req):
-	if req.method == "GET":	
-		sortBy = req.GET.get('sortBy','-startDate')
-		count = req.GET.get('count',None)
+def allEvents(request):
+	if request.method == "GET":	
+		sortBy = request.GET.get('sortBy','-startDate')
+		count = request.GET.get('count',None)
 	
 		# How Many Events to Return
 		if count != None:
@@ -31,74 +22,76 @@ def events(req):
 		
 		return HttpResponse(eventsList, content_type="application/json")
 
-			
-def add(req):
-	if req.method == "GET":
+def event(request, id):
+	if request.method == "GET":	
+		event = Event.objects.filter(id=id)
+		
+		if len(event) > 0:
+			return HttpResponse(event[0], content_type="application/json")
+
+def add(request, orgID):
+	if request.method == "GET":
 		form = EventForm()
 		
-		csrf_token = get_token(req)
+		csrf_token = get_token(request)
 		csrf_token_html = '<input type="hidden" name="csrfmiddlewaretoken" value="{}" />'.format(csrf_token)
 
 		return HttpResponse("<form method='post'>" + form.as_ul() + csrf_token_html + "<input type='submit'></form>")
 		
-	elif req.method == "POST":
+	elif request.method == "POST":
 		# TODO: Do some validation
-		form = EventForm(req.POST)
+		form = EventForm(request.POST)
 	
 		if form.is_valid():
-			title = form.cleaned_data["title"]
-			location = form.cleaned_data["location"]
-			description = form.cleaned_data["description"]
-			startDate = form.cleaned_data["startDate"]
-			endDate = form.cleaned_data["endDate"]
-			tags = form.cleaned_data["tags"]
-			owner = form.cleaned_data["owner"]
-			capacity = form.cleaned_data["capacity"]
+			organization = Organization.objects.filter(id=orgID)
 			
-			newEvent = Event.objects.create(
-				title = title,
-				location = location,
-				description = description,
-				startDate = startDate,
-				endDate = endDate,
-				tags = tags,
-				owner = owner,
-				capacity = capacity
-			)
+			if len(organization) > 0:
+				organization = organization[0]
+				
+				if organization in request.user.student.organizations.all():
+					event = form.save()
+					event.owner = request.user.username
+					event.save()
+					event.organizations.add(organization)
 			
-			return HttpResponse("Success")
-		else:
-			return HttpResponse("Fail")
+					return HttpResponse("Success")
+
+	return HttpResponse("Fail")
     
-def update(req, id):
+def update(request, id):
 	updatingEvent = Event.objects.filter(id=id)
 				
 	if len(updatingEvent) > 0:
 		updatingEvent = updatingEvent[0]
 		
-		if req.method == "GET":		
+		if request.method == "GET":		
 			form = EventForm(instance=updatingEvent)
 		
-			csrf_token = get_token(req)
+			csrf_token = get_token(request)
 			csrf_token_html = '<input type="hidden" name="csrfmiddlewaretoken" value="{}" />'.format(csrf_token)
 
 			return HttpResponse("<form method='post'>" + form.as_ul() + csrf_token_html + "<input type='submit'></form>")
 			
-		elif req.method == "POST":
-			form = EventForm(req.POST, instance=updatingEvent)
-			form.save()
-			return HttpResponse("Success")
+		elif request.method == "POST":
+			form = EventForm(request.POST, instance=updatingEvent)
+			
+			for organization in form.organizations.all():
+				if organization in request.user.student.organizations.all():
+					form.save()
+					return HttpResponse("Success")
 
 	return HttpResponse("No Event Found")
 
 
-def delete(req, id):
+def delete(request, id):
 	# TODO: Do Some validation	
 	deletingEvent = Event.objects.filter(id=id)
 	
 	if len(deletingEvent) > 0:
-		deletingEvent.delete()
-		return HttpResponse("Success")
+		for organization in deletingEvent[0].organizations.all():
+			if organization in request.user.student.organizations.all():
+				deletingEvent.delete()
+				return HttpResponse("Success")
 	
 	return HttpResponse("Fail")
